requirements.txt
langchain==0.3.25
openai==1.88.0
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pytest==8.2.1
langchain-openai==0.3.23
faiss-cpu==1.11.0
langchain-community
langchain-core==0.3.65
